+++ src/components/SrOnly/index.ts
@@ -2,8 +2,8 @@ import styled from "styled-components";
 
 export const SrOnly = styled.input`
   position: absolute;
-  width: 1px;
-  height: 1px;
+  width: 0.0625rem;
+  height: 0.0625rem;
   overflow: hidden;
   clip: rect(0 0 0 0);
   clip-path: inset(50%);